# packages/frontend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "德州扑克游戏API",
    "version": "1.0.0",
    "description": "德州扑克游戏后端API文档"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/social/friend-requests": {
      "post": {
        "summary": "发送好友请求",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "toUserId": {
                    "type": "string",
                    "description": "接收者用户ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "好友请求发送成功"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/friend-requests/{id}/respond": {
      "post": {
        "summary": "响应好友请求",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "好友请求ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "accepted",
                      "rejected"
                    ],
                    "description": "响应状态"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "响应成功"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/friend-requests/received": {
      "get": {
        "summary": "获取收到的好友请求",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "accepted",
                "rejected"
              ]
            },
            "description": "请求状态"
          }
        ],
        "responses": {
          "200": {
            "description": "好友请求列表"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/friend-requests/sent": {
      "get": {
        "summary": "获取发送的好友请求",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "accepted",
                "rejected"
              ]
            },
            "description": "请求状态"
          }
        ],
        "responses": {
          "200": {
            "description": "好友请求列表"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/friend-requests/{id}": {
      "delete": {
        "summary": "删除好友请求",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "好友请求ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/friends": {
      "get": {
        "summary": "获取好友列表",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "好友列表"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/friends/{friendId}": {
      "delete": {
        "summary": "删除好友",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "friendId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "好友ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/friends/check/{friendId}": {
      "get": {
        "summary": "检查是否是好友",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "friendId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "好友ID"
          }
        ],
        "responses": {
          "200": {
            "description": "检查结果"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/messages": {
      "post": {
        "summary": "发送私信",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "receiverId": {
                    "type": "string",
                    "description": "接收者ID"
                  },
                  "content": {
                    "type": "string",
                    "description": "消息内容"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "消息发送成功"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/messages/{otherUserId}": {
      "get": {
        "summary": "获取与某个用户的私信历史",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "otherUserId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "对方用户ID"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "number"
            },
            "description": "限制数量"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "number"
            },
            "description": "偏移量"
          }
        ],
        "responses": {
          "200": {
            "description": "私信历史列表"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/messages/unread-count": {
      "get": {
        "summary": "获取未读私信数量",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "未读消息数量"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/messages/read": {
      "post": {
        "summary": "标记私信为已读",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messageIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "消息ID列表"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "标记成功"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/social/messages/{otherUserId}/read-all": {
      "post": {
        "summary": "标记与某个用户的所有私信为已读",
        "tags": [
          "Social"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "otherUserId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "对方用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "标记成功"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
  },
...
}